Linux: add support for tm_zone to localtime via the sandbox.
Glibc adds a field to struct tm, tm_zone, which contains a pointer to
a string for the current time zone. Since this makes struct tm a
non-POD data type, it only works for the non-thread-safe localtime()
call, not localtime_r().

However, V8 relies on it so this patch adds support for passing it
through our sandbox.

BUG=17458
http://codereview.chromium.org/159216


git-svn-id: http://src.chromium.org/svn/trunk/src@21311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c