python 2.4's implementation of sys.popen() has a race condition
python 2.4's implementation of sys.popen() keeps a list of subprocesses
around to handle cleanup, but the list isn't safe against concurrent access,
leading it to sometimes through a ValueError exception. The list is kind
of pointless, and the exception can be ignored (which it is in 2.5).

BUG=none
R=maruel
TEST=none

Review URL: http://codereview.chromium.org/155782

git-svn-id: http://src.chromium.org/svn/trunk/src@21367 4ff67af0-8c30-449e-8e8b-ad334ec8d88c