Linux: splice together RendererPreferences and SkFontHost
Dan Erat has been working down, plumbing the user's font settings from
XSETTINGS in the UI thread of the browser to RendererPreferences in
the renderer. At the same time, I've been working upwards:
implementing subpixel rendering support in Skia and adding an
interface to it in WebCore.

This patch joins the two together. It adds an API surface to the
WebKit API for configuring the font preferences and adds code to
translate from RendererPreferences to that API. It's a little
unfortunate that we need four sets of enums around the code base for
things like the hinting level, but each is supposed to be self
contained and adding a dependancy would much more grevious.

BUG=12179
http://codereview.chromium.org/159152


git-svn-id: http://src.chromium.org/svn/trunk/src@21493 4ff67af0-8c30-449e-8e8b-ad334ec8d88c