Paste at the beginning of a middle click rather than after letting the page handle it.
The problem is that the page changes the focused area in the mouse release handler.

Even after this fix, it is possible to paste into the wrong place. The place that gets the text is whichever text input is focused when the browser sends back the clipboard contents, not the text input that is focused when the middle click is handled. This has always been true but seems harmless in most cases.

BUG=17504
TEST=open gmail chat. Put something in your PRIMARY. Click in the text entry area. Middle click in the chat display area. The selection shouldn't paste into the text entry. Otherwise, middle click paste should still work.
TEST=the reduction in the bug acts as expected
Review URL: http://codereview.chromium.org/160065

git-svn-id: http://src.chromium.org/svn/trunk/src@21553 4ff67af0-8c30-449e-8e8b-ad334ec8d88c