Linux: add GYP variable to use the system libpng.
Since our packagers have been requesting that we support using more
system libraries, I'm starting to commit these sorts of changes.
However, please note that one MUST NOT use the system libpng on, at
least, Ubuntu 64-bit. Because Chromium on Linux is still a 32-bit
build we end up with the 32-bit compat versions of libraries and, on
Ubuntu at least, this includes a very old libpng with security issues.


git-svn-id: http://src.chromium.org/svn/trunk/src@21663 4ff67af0-8c30-449e-8e8b-ad334ec8d88c