Move window to workarea origin if it is completely off-screen.
BUG=17822
TEST=Use multiple monitors. Start Chrome, move the window to a non-primary
monitor, and exit Chrome. Disconnect or disable the monitor and start Chrome
again. Chrome should appear at the origin of a remaining monitor. Also try
chaning the resolution of the non-primary monitor. As far as the window fits
within the monitor, it should be shown as-is. If not, it is moved to the origin
and resized, if necessary.

Review URL: http://codereview.chromium.org/160246

git-svn-id: http://src.chromium.org/svn/trunk/src@21946 4ff67af0-8c30-449e-8e8b-ad334ec8d88c