Retry requests that has timed out in BufferedDataSource
Retries request if the read request is not fulfilled in 5 seconds and at most 3 times. This will handle the case when
a user pause a movie and starts it in a later time when the
connection to the server has already timed out.

TEST=BufferedDataSourceTest.*
BUG=17479

done

Review URL: http://codereview.chromium.org/159382

git-svn-id: http://src.chromium.org/svn/trunk/src@21982 4ff67af0-8c30-449e-8e8b-ad334ec8d88c