Update the Mac plugin dummy window code for the new coordinate system.
Fixes the location of the dummy window, cleans up some duplicate window-handling code, and tries to keep the window activation status more in sync with what it would be in a single-process system. The net result is that event handling in plugins works better, although there are still activation issues.

BUG=none
TEST=Clicks on plugins should register in the right place (assuming they register at all; there are still bugs in having them handled).

Review URL: http://codereview.chromium.org/159525

git-svn-id: http://src.chromium.org/svn/trunk/src@21987 4ff67af0-8c30-449e-8e8b-ad334ec8d88c