Implement first cut at the extension installation prompt on Windows.
Had to move set_path() earlier in the unpack process so that we would have real paths for the resources, so that we could load and display the icon in the ui.

This exposed a problem where page actions had absolute paths and other images in extensions had relative paths. Extension::GetBrowserImages() was expecting relative paths, and it just happened to work because in both cases Extension was initialized without a path.

Modified page actions to use relative paths to be consistent with other images.

Review URL: http://codereview.chromium.org/160516

git-svn-id: http://src.chromium.org/svn/trunk/src@22368 4ff67af0-8c30-449e-8e8b-ad334ec8d88c