linux: fix windowless Flash when scrolling
Apparently on linux windowless Flash needs a paint after a NPP_SetWindow, otherwise it cancels its timer to do InvalidateRect.

Also in this change: WebKit calls WebPluginImpl::setFrameRect at each paint, even if the rects haven't changed. That used to always send a message to the plugin process, which is unnecessary (that case is always a no-op in the WebPluginDelegateImpl), so I fixed that too.

BUG=18423

Review URL: http://codereview.chromium.org/159907


git-svn-id: http://src.chromium.org/svn/trunk/src@22702 4ff67af0-8c30-449e-8e8b-ad334ec8d88c