Reverts the revert of 22517.
Rewrites the Mac view resizing logic to have the BrowserWindowController
directly resize and relayout its children views.  Now when a view needs
to be resized, it asks its resize delegate (typically its controller's
parent) to perform the resize.

BUG=http://crbug.com/17619

TEST=Make sure that views are laid out correctly, even when they change size.
Open and close the bookmark bar.  Trigger an infobar and then close it.  Trigger
the download shelf and then close it.  Trigger a download shelf with the infobar
open, or with the bookmark bar open.  Switch to and from fullscreen with various
bars open.  Resize the browser window with various bars open.  Start the browser
with and without the bookmark bar open.  Try all of the above in a popup window.

Review URL: http://codereview.chromium.org/165014

git-svn-id: http://src.chromium.org/svn/trunk/src@22721 4ff67af0-8c30-449e-8e8b-ad334ec8d88c