Add short circuit case for self assignment, and add virtual destructor
Pickle is a base class of Message, which has a virtual destructor, so
it is better to have Pickle have a virtual destructor as well.

Add more defensive code in the case of self assignment.
**IF** self assignment was used, and the target was read-only,
then the existing code would have a memory fault.
I don't think this is currently possible in use cases, but
the defensize code is much better. (per suggestion from CPU
on last CL).

BUG=17027
BUG=17088
r=cpu
Review URL: http://codereview.chromium.org/159800

git-svn-id: http://src.chromium.org/svn/trunk/src@22826 4ff67af0-8c30-449e-8e8b-ad334ec8d88c