Lands http://codereview.chromium.org/159838 for Pierre-Antoine. From him:
Right clicking the back and forward buttons is buggy. A single right click opens
the drop down menu, however the button does not enter the PUSHED state. If you
double right click then click on the tab content area, the button stays in a
PUSHED state until you mouseover it.

The fix should replicate the intuitive behavior. On a single right click, the
button enters the PUSHED state and the drop down appears immediately. As soon as
the user clicks on another View, the drop down closes and the button returns to
the NORMAL state. Double right click behaves similarly. 

BUG=9717
TEST=none

Review URL: http://codereview.chromium.org/165318

git-svn-id: http://src.chromium.org/svn/trunk/src@23101 4ff67af0-8c30-449e-8e8b-ad334ec8d88c