Don't focus the location bar if we're in full screen mode.
I originally tried to set the GTK_CAN_FOCUS flag on the autocomplete
edit text_view, but that requires plumbing a method through BrowserToolbar,
LocationBarGtk, and AutocompleteEditViewGtk.  This is easier.

Since the widget is already hidden, it's not in the tab order so other
than explicit calls in code, it's not possible to give the gtk text view
focus.

BUG=14945

Review URL: http://codereview.chromium.org/164342


git-svn-id: http://src.chromium.org/svn/trunk/src@23105 4ff67af0-8c30-449e-8e8b-ad334ec8d88c