Rebaseline block-test* tests for Mac.
On Mac, when trying to connect to 255.255.255.255, we get an EACCES error.  I
don't think it is correct for me to map that to ERR_ADDRESS_INVALID or
ERR_ADDRESS_UNREACHABLE, so I'm going to map it to ERR_ACCESS_DENIED.  This
means that I need to rebaseline the chromium-mac results for these tests.

I think it is okay to have platform specific behavior here.  I kind of prefer
that the error codes we show are more closely in sync with the operating
system.  That way it may be easier for a sophisticated user to make sense of
our error codes.

R=wtc
BUG=18665
TEST=none

Review URL: http://codereview.chromium.org/164114

git-svn-id: http://src.chromium.org/svn/trunk/src@23176 4ff67af0-8c30-449e-8e8b-ad334ec8d88c