Linux sandbox: fix security issue.
(Reported by Julien Tinnes)

Because the chroot helper process and the zygote share a FILES
structure, the latter can race the former and change the value of cwd
before it does chroot("."). Because of this, the zygote could chroot
into a directory of its choosing. Once there, it could setup hardlinks
to SUID binaries and possibly make them misbehave if they weren't
sufficiently paranoid.

This possibility should have been migigated by the removal of
dangerous environment variables. However, we had to reinstate them in
order to pass LD_LIBRARY_PATH because some setups don't have ld.so
setup to use /usr/lib32 and also for ffmpeg.

http://codereview.chromium.org/164427


git-svn-id: http://src.chromium.org/svn/trunk/src@23228 4ff67af0-8c30-449e-8e8b-ad334ec8d88c