Defer closing the TestShell window.
This change causes us to match the behavior of Windows and Linux, and it also
better conforms to the API (closeWidgetSoon) which does not expect things to be
synchronously closed.

Note: I do not believe it is important to null out shell_, so I deleted that
line.  The other platforms are similar.

R=smorgan
BUG=19205
TEST=covered by existing layout tests

Review URL: http://codereview.chromium.org/164431

git-svn-id: http://src.chromium.org/svn/trunk/src@23244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c