Refcounting BufferedResourceLoader
BUG=18677

We do range request for <video> and <audio>, we used
to destroy the ResourceLoaderBridge right after we
cancel the request. But it turns out that this would
leave an excessive amount of warnings in the renderer
and also leaking of URLRequest in the browser.

By refcounting BufferedResourceLoader (the container
of ResourceLoaderBridge) that does resource loading
for <video>, we can then destroy the
ResourceLoaderBridge after request has completed.

Review URL: http://codereview.chromium.org/164361

git-svn-id: http://src.chromium.org/svn/trunk/src@23274 4ff67af0-8c30-449e-8e8b-ad334ec8d88c