Move glue/event_conversion.{h,cc} into the WebKit API implementation.
I also renamed the MakeFoo classes to FooBuilder so that it would be
more clear at the call sites that you are dealing with a class.

I then translated the ToBar functions to also be BarBuilder classes.
I just did this for consistency with the FooBuilders.

For reference, "Foo" refers to WebCore Platform*Event classes and
"Bar" refers to WebInputEvent subclasses :-)

R=jorlow
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/164524

git-svn-id: http://src.chromium.org/svn/trunk/src@23425 4ff67af0-8c30-449e-8e8b-ad334ec8d88c