Fix typo from r23272.
Also change NavigateToURLLogResult to take a std::string for its URL argument.

Also prepend g_ on all the file-global static variables. This should help prevent future mistakes of a similar nature. This is not really within style but sort of matches the g_ for global variables. Anyway this file is pretty far out of style to begin with.

Review URL: http://codereview.chromium.org/169009

git-svn-id: http://src.chromium.org/svn/trunk/src@23484 4ff67af0-8c30-449e-8e8b-ad334ec8d88c