plugins: unfork WebPluginDelegateImpl between linux/win
Algorithm:
 - copy win to shared
 - for each diff between shared and lin, delete/refactor shared
 - for each diff between shared and lin, delete in lin
 - for each diff between shared and win, delete in win

The change should be mostly deletes.
Please verify that I didn't change any behavior on Windows or GTK.

Note: coincidentally, this change also fixes a bug that we've seen
on Linux that was fixed in the Windows version of the file.  I only
noticed it because the Windows/Linux files differed, but after this
unforking the change is now in the shared code.

Mac shouldn't be too hard to do too.

Review URL: http://codereview.chromium.org/174479

git-svn-id: http://src.chromium.org/svn/trunk/src@24404 4ff67af0-8c30-449e-8e8b-ad334ec8d88c