Fix two issues with the plugin installer.
The first is that we weren't reloading the pages with plugins.

The second is that there was a InstallerDialog instance per plugin object on the page.  On a page like nytimes.com that creates and destroys a Flash object on loading, this meant that if the user clicked on the plugin/infobar inbetween, the dialog would show up and then get hidden abruptly.

TEST=go to nytimes.com on a machine without Flash installed and click the infobar as soon as it shows up.
BUG=20690
Review URL: http://codereview.chromium.org/179051

git-svn-id: http://src.chromium.org/svn/trunk/src@24995 4ff67af0-8c30-449e-8e8b-ad334ec8d88c