Reduce the size of the paint queue.
(Sheriffs: warning, dangerous patch)

On Chronos, with the slower CPU, it appears that page scrolling
sigificantly lags the mouse cursor sometimes.  Some amount of lag will
always happen, but it appears that we are queuing old mouse positions.

This patch makes it so that we don't ACK a mouse move event if there
is a resulting paint pending. Current we get something like this:

 Renderer:                                 Browser:
           <------- HandleInputEvent --------
 (handles the event and defers
  the painting)
           --------   ACK ------------------->
                                          (has a mouse move saved up)
               ... HandleInputEvent ---------
 (painting starts)
          ----------- PaintRect ------------->
          <----...

At this point, we buffered a mouse position while painting. If we
don't ACK the event until we have finished painting, we can get a more
up-to-date event.

This will need careful observation of the perf bots to check that
nothing bad happens.

http://codereview.chromium.org/173285


git-svn-id: http://src.chromium.org/svn/trunk/src@25059 4ff67af0-8c30-449e-8e8b-ad334ec8d88c