Simplify the gtk location bar padding and make it more consistent.
Sets the container border and box padding values, so that each widget doesn't need an alignment to pad itself.

With this the horizontal padding is almost the same as windows (in a few cases we have 1 more pixel between elements.)
The vertical padding should be unchanged.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/195048

git-svn-id: http://src.chromium.org/svn/trunk/src@25831 4ff67af0-8c30-449e-8e8b-ad334ec8d88c