Add a bare-bones extension shelf that displays extension items on OS X.
This brings our extension support to about the level it has on linux.

One issue is that the toolstrips are webpages with a background image that just happens to look like the shelf they are on. But the background images are not updated on key->nonkey window changes, so the toolstrip backgrounds look slightly off in one of the two cases. If we decide to keep the shelf, we should fix this, but see the bug for erikkay's stance on this.

Also, the NTP is only loaded after all toolstrips have been loaded for some reason. That's what happens on the other platforms too, I believe.

The extension shelf uses the DownloadShelfView as background view for now.

Screenie: http://imgur.com/wSHgU.png

BUG=19073
TEST=Extensions that live in the shelf should show up. They should be clickable, resize correctly (e.g. the build status extension), and the shelf should interact in a sane way with the status bubble.

Review URL: http://codereview.chromium.org/175025

git-svn-id: http://src.chromium.org/svn/trunk/src@26311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c