Add Profile Import Process Type.
This is currently only used on OS X.

We need a separate process to do profile import for 3 reasons:
1)In Firefox we load external libraries which we have no control over.
2)Conceptually it's safer to run parsing code OOP.
3)quoth chrome/browser/importer/nss_decryptor_mac.mm :
"""
On OS X the nss3 libraries are compiled with depedencies
on one another, referenced using dyld's @executable_path directive.
To make a long story short in order to get the libraries to load, dyld's
fallback path needs to be set to the directory containing the libraries.
To do so, the process this function runs in must have the
DYLD_FALLBACK_LIBRARY_PATH set on startup to said directory.
"""

BUG=14458
TEST=None Possible

Review URL: http://codereview.chromium.org/193131

git-svn-id: http://src.chromium.org/svn/trunk/src@26501 4ff67af0-8c30-449e-8e8b-ad334ec8d88c