While fixing bug 22070 (Toolstrip theme background not updated when resizing Chrome window) I added code to trigger repainting the background on resize. This flushed out a problem (Skia exceptions) that was causing the extension tests to become flaky on the bots (bug 22135). I've changed the code to do a deep copy on the subset bitmap we get before passing it down to the renderer, which causes Skia to not go haywire when drawing the background.
That makes it possible to fix bug 22070, which now causes the background to be updated on resize. There is also now no need to set the background_needs_repaint_ flag to true before calling Layout. We also don't need to keep track of whether we are detached or not, we just need to monitor the background_needs_repaint_ flag.

BUG=22070, 22135
TEST=Monitor the bots and see if the tests become flaky again.
Review URL: http://codereview.chromium.org/215017

git-svn-id: http://src.chromium.org/svn/trunk/src@26526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c