Forward edit commands to be executed for shortcuts which aren't handled by javascript.
As a result, cocoa keybindings work.

The approach:

We instantiate a vanilla NSResponder, call interpretKeyEvents on it, and
record all of the selectors passed into doCommandBySelector while
interpreting the key event.  The selectors are converted into edit commands
which can be passed to the render process.

Caveats:
- Shortcuts involving a sequence of key combinations (chords) don't work,
because we instantiate a new responder for each event.
- We ignore key combinations that don't include a modifier (ctrl, cmd, alt)
because this was causing strange behavior (e.g. tab always inserted a tab
rather than moving to the next field on the page).

BUG=12538
TEST=go to a page with a text field. make sure that ctrl-a moves cursor to beginning of line, ctrl-e goes to the end, ctrl-d deletes to the right, ctrl-h to the left, etc

Review URL: http://codereview.chromium.org/209046
Patch from matt@tolton.com.

git-svn-id: http://src.chromium.org/svn/trunk/src@26694 4ff67af0-8c30-449e-8e8b-ad334ec8d88c