Add Apple's ImageAndTextCell to third_party.
I asked dannyb@, he was ok with that as long as we put it somewhere into third_party. The code will be used e.g. in the search engine manager, the task manager and maybe in the frozen renderer dialog.

I copied the files from /Developer/Examples/AppKit/DragNDropOutlineView on a Leopard system (and copied the license notice from the m file to the h file).

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/231026

git-svn-id: http://src.chromium.org/svn/trunk/src@27211 4ff67af0-8c30-449e-8e8b-ad334ec8d88c