Add the ability to distinguish between tests that fail due to a difference in the text output, tests that fail due to differences in the image output, and tests that fail both.
This adds the additional failure types 'IMAGE', 'TEXT', and 'BOTH' to 
test_expectations.txt. Tests that are marked 'FAIL' should eventually be
migrated to one of the above three, but for now FAIL means that any of the
three types might happen. 

Note that we do not distinguish between a text diff and a simplified text
diff; 'TEXT' will cover either or both types of failures.

  BUG=none
  TEST=none
  R=ojan@chromium.org

Review URL: http://codereview.chromium.org/235016

git-svn-id: http://src.chromium.org/svn/trunk/src@27305 4ff67af0-8c30-449e-8e8b-ad334ec8d88c