More work in Command Buffers
I think/hope this is the last big CL for this.

*) Moved GAPIInterface enums to cmd_buffer_format.h

The reason is you should not have to include CODE 
to make data. The fact that the enums were in 
GAPIInterface, a class, meant you had to include
code if you wanted to make a command buffer.

*) Typed arguments where I could.

So for example Draw takes PrimitiveType instead of
uint32. CreateVertexBuffer takes a ResourceId

*) Renamed enums etc to match style guide.

*) Moved BitFields into specific commands

*) renamed arguments from id to type_id.



Review URL: http://codereview.chromium.org/234002

git-svn-id: http://src.chromium.org/svn/trunk/src@27322 4ff67af0-8c30-449e-8e8b-ad334ec8d88c