Linux: Make InfoBubble use an override-redirect (popup) window.
This makes it work correctly in ion3 and other window managers that don't expect clients to try to move top-level windows themselves.

This implementation grabs the pointer and keyboard.  By doing this and using an override-redirect window, we should be able to avoid worrying about interactions with different window managers.  The only downside (alluded to in the previous code) is that window manager keybindings don't make it through until the user dismisses the bubble by clicking outside of it or hitting Enter or Escape.  I don't think this will be a problem; it's no different from what happens when you open a context menu in an app.

BUG=20523
TEST=tested first-run and bookmark bubbles in Metacity, ion3, Fluxbox, KDE4, and the WM that I'm working on

Review URL: http://codereview.chromium.org/198016

git-svn-id: http://src.chromium.org/svn/trunk/src@27578 4ff67af0-8c30-449e-8e8b-ad334ec8d88c