Remove all MemoryModel-related code, since we're unlikely to use it again.  Also changes some sample code to use a random generic name ("PruningAlgorithm") instead of MemoryModel in hopes that no one will be confused.
This does not remove IdleTimer, even though it now has no users.  I do plan to remove it, but I want to do that in a separate commit so it's easy to reference and revert if at some later point we want this functionality.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/242079

git-svn-id: http://src.chromium.org/svn/trunk/src@27652 4ff67af0-8c30-449e-8e8b-ad334ec8d88c