Linux: Improve info bubble placement.
Make InfoBubble's constructor take positions relative to the browser
window's origin, rather than the absolute positions used by Views.

Otherwise, we run into issues when the browser window is being moved while
the bubble is created -- this is unlikely to happen for bookmark bubbles,
but when the first run bubble is created, the browser window may or may not
be in the final place that the window manager is going to put it.  There
was previously a 200 ms delay before opening the bubble, but that approach
still fails sometimes when X is being forwarded over a network connection.

This change makes us update the bubble's position as we receive
ConfigureNotify events about its browser window.

TEST=checked that first-run and bookmark bubbles are placed correctly in openbox, KDE, metacity, fluxbox, and ion3

Review URL: http://codereview.chromium.org/256068

git-svn-id: http://src.chromium.org/svn/trunk/src@28232 4ff67af0-8c30-449e-8e8b-ad334ec8d88c