Coverity: fix possible leak in VisitSegmentDatabase.
If we get a non-zero segment_id, a zero segment_id, and another non-zero segment_id, the pud for the zero segment_id will be leaked.  A non-zero segment_id shouldn't happen since it refers to a primary-key in the segments table, but this changes it to check directly whether a pud has been created, and thus makes it more error-proof.  This should be safe, as pud is only stored into the list in two places, in the first it is immediately replaced with a new pud, in the second it is storing the final pud (if any.)

CID=6379
TEST=none
BUG=none

Review URL: http://codereview.chromium.org/265027

git-svn-id: http://src.chromium.org/svn/trunk/src@28445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c