This fixes the rendering issue in Firefox by ignoring zero height resize requests from Firefox.
When the fullscreen stuff was redone (r26489), we removed a
WM_SIZE handler that would call the plugin's "Resize" API, and
moved it into NPP_SetWindow, but at the same time a SetWindowPos
call was also added that would take the window size as gospel and
set it, but only if we're not in fullscreen mode.  Turns out that
this is not necessary because the Resize API already checks to
see if we're in full-screen mode, and handles the spurious zero
height messages properly too.  So I just took out the explicit
SetWindowPos.
Review URL: http://codereview.chromium.org/268022

git-svn-id: http://src.chromium.org/svn/trunk/src@28482 4ff67af0-8c30-449e-8e8b-ad334ec8d88c