Revert "Revert "Start the database query for the new tab page results as soon""
This re-applies r28549 with a NULL pointer check for the HistoryService, which
can be NULL during unit tests.

It also means that if the user does a view-source on the new tab page, we make
a wasted query to the history service, but I guess that's not a big deal.

TBR=arv

Review URL: http://codereview.chromium.org/267039

git-svn-id: http://src.chromium.org/svn/trunk/src@28559 4ff67af0-8c30-449e-8e8b-ad334ec8d88c