Implement a version of NativeMetafile for the Mac
This is the first of series of patches to implement printing on the Mac.

This API is based heavily on the Linux version; although it doesn't end up abstracting very much on the Mac, having a NativeMetafile implementation makes it easy to pass data through the existing print architecture.

BUG=13158
TEST=Printing unit tests; no effect in the application yet.

Review URL: http://codereview.chromium.org/269041

git-svn-id: http://src.chromium.org/svn/trunk/src@28722 4ff67af0-8c30-449e-8e8b-ad334ec8d88c