Add env variable to write minidumps to a preset directory.
Add an |BREAKPAD_DUMP_LOCATION| environmental variable to allow overriding the default location
for minidump generation.

This is for ChromeBot who need a single copy of the Chrome bundle which can be run several times
alternating the location of crash dumps.

This isn't implemented as a command line flag because only some command line flags are passed to child processes and we need a robust solution.

TEST=Minidumps should still be written for browser and renderer crashes.

Review URL: http://codereview.chromium.org/264038

git-svn-id: http://src.chromium.org/svn/trunk/src@29126 4ff67af0-8c30-449e-8e8b-ad334ec8d88c