Within libjingle, add an abstraction of winsock initialization.
When building chrome, override the default implementation so
that we share the winsock initialization singleton with net/base.

Trigger the winsock initialization lazily in a few places.

This is intended to fix a startup perf regression that appeared
when sync (and libjingle) became linked into chrome.dll, rather
than a separate delay-loaded sync dll.

BUG=24448
TEST=Verified deferred winsock initialization via debugger.  Measured before and after time of the StartupTest.PerfTest. Verified sync notification still work.

Review URL: http://codereview.chromium.org/267089

git-svn-id: http://src.chromium.org/svn/trunk/src@29228 4ff67af0-8c30-449e-8e8b-ad334ec8d88c